// lib/gen/models/GetUserResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MVC
 * MVC Assignment
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { Role } from './Role';
import { RoleFromJSON, RoleToJSON } from './Role';

/**
 *
 * @export
 * @interface GetUserResponse
 */
export interface GetUserResponse {
  /**
   *
   * @type {string}
   * @memberof GetUserResponse
   */
  email?: string;
  /**
   *
   * @type {number}
   * @memberof GetUserResponse
   */
  id?: number;
  /**
   *
   * @type {string}
   * @memberof GetUserResponse
   */
  name?: string;
  /**
   *
   * @type {Role}
   * @memberof GetUserResponse
   */
  role?: Role;
}

/**
 * Check if a given object implements the GetUserResponse interface.
 */
export function instanceOfGetUserResponse(value: object): value is GetUserResponse {
  return true;
}

export function GetUserResponseFromJSON(json: any): GetUserResponse {
  return GetUserResponseFromJSONTyped(json, false);
}

export function GetUserResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetUserResponse {
  if (json == null) {
    return json;
  }
  return {
    email: json['email'] == null ? undefined : json['email'],
    id: json['id'] == null ? undefined : json['id'],
    name: json['name'] == null ? undefined : json['name'],
    role: json['role'] == null ? undefined : RoleFromJSON(json['role'])
  };
}

export function GetUserResponseToJSON(json: any): GetUserResponse {
  return GetUserResponseToJSONTyped(json, false);
}

export function GetUserResponseToJSONTyped(value?: GetUserResponse | null, ignoreDiscriminator: boolean = false): any {
  if (value == null) {
    return value;
  }

  return {
    email: value['email'],
    id: value['id'],
    name: value['name'],
    role: RoleToJSON(value['role'])
  };
}